requirements.txt
requests
redis
psycopg2
pymongo
newspaper3k
timeout_decorator
tweepy
numpy
praw
facebook-sdk
jellyfish
rq
pycountry
raven
python-decouple
python-crontab
tornado
jinja2
bson